Primera actualización con GPU
